% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RBF.R
\name{plot.ReplicationBF}
\alias{plot.ReplicationBF}
\title{Plotting \code{ReplicationBF} objects}
\usage{
plot.ReplicationBF(x, ..., use.ggplot = TRUE)
}
\arguments{
\item{use.ggplot}{If \code{TRUE} uses \code{ggplot2}. \code{FALSE} not yet
supported.}

\item{rbf.object}{Object from an \code{RBF_*} call. Has to include Posterior
samples (\code{store.samples = TRUE}).}
}
\value{
A \code{ggplot2} object containing the plot.
}
\description{
This method can be used to visually represent a Replication Bayes Factor.
}
\details{
\code{plot.ReplicationBF} generally is a Savage-Dickey ratio representation:
Prior and Posterior distributions are plotted and the ratio of the
distributions at \eqn{\delta = 0} is calculated (see Wagenmakers et al.,
2010).
}
\examples{
\dontrun{
# Run RBF function with store.samples = T to save posterior samples used
# for plotting
rbf <- RBF_ttest(2.18, c(10, 11), 3.06, c(27, 27), method = "NormApprox",
  store.samples = T)
plot(rbf)

# If samples are missing, plot() will throw an error:
rbf <- RBF_ttest(2.18, c(10, 11), 3.06, c(27, 27), method = "NormApprox")
plot(rbf)
}

}
\references{
\itemize{
    \item \insertRef{Wagenmakers2010}{ReplicationBF}
}
}
