% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RBF.R
\name{plot_RBF}
\alias{plot_RBF}
\title{Plot the results of a Replication Bayes Factor.}
\usage{
plot_RBF(rbf.object)
}
\arguments{
\item{rbf.object}{Result from \code{\link{RBF_ttest}} or
\code{\link{RBF_Ftest}}.}
}
\value{
None.
}
\description{
Plot the results of a Replication Bayes Factor.
}
