% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF_corr.R
\name{RBF_corr}
\alias{RBF_corr}
\title{Calculate a Replication Bayes Factors for Correlations.}
\usage{
RBF_corr(r.orig, n.orig, r.rep, n.rep, M = 1e+06)
}
\arguments{
\item{r.orig}{Correlation coefficient of the original study.}

\item{n.orig}{Number of observations in the original study.}

\item{r.rep}{Correlation coefficient in the replication study.}

\item{n.rep}{Number of observaions in the replication study.}

\item{M}{Number of MCMC draws to approximate the original study's posterior.}
}
\value{
(none), because the method is not yet implemented.
}
\description{
Calculate a Replication Bayes Factors for Correlations.
}
