% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF_Ftest.R
\name{RBF_Ftest}
\alias{RBF_Ftest}
\title{Calculate a Replication Bayes Factor for F-Tests.}
\usage{
RBF_Ftest(F.orig, df.orig, N.orig, F.rep, df.rep, N.rep, M = 1e+05,
  store.samples = FALSE)
}
\arguments{
\item{F.orig}{F-statistic from the original study.}

\item{df.orig}{Numeric vector containing the degrees of freedom for the
F-test from the original study.}

\item{N.orig}{Total number of observations in the original study.}

\item{F.rep}{F-statistic from the replication study.}

\item{df.rep}{Numeric vector containing the degrees of freedom for the F-Test
from the replication study.}

\item{N.rep}{Total number of observations in the replication study.}

\item{M}{Number of draws from the posterior distribution to approximate the
marginal likelihood.}

\item{store.samples}{If \emph{TRUE}, the samples of the original's posterior
distribution are stored in the return object.}
}
\value{
An \code{ReplicationBF} object containing the value of the
    Replication Bayes Factor in \code{bayesFactor}.
}
\description{
\code{RBF_Ftest} calculates a Replication Bayes Factor for F-Tests from
balanced fixed effect, between subject ANOVA designs.
}
\details{
The Replication Bayes Factor is a marginal likelihood ratio between two
positions (see Verhagen & Wagenmakers, 2014):
\itemize{
    \item \strong{H0}: The position of a skeptic, who does not place
        confidence in the findings of the original study and assumes
        \eqn{\theta \approx 0}.
    \item \strong{Hr}: The position of a proponent, who expects the original
        study to be a faithful estimation of the effect size and for which
        the replication offers additional evidence. This is formulated as
        \eqn{\theta \approx \theta_{orig}} with \eqn{\theta_{orig}} being
        the effect size estimate from the original study.
}

The Bayes factor is estimated through Importance Sampling (Gamerman & Lopes,
2006). The importance density is half-normal with parameters estimated from
the posterior distribution. Posterior distribution is sampled using
Metropolis-Hastings from \code{MCMCpack::MCMCmetrop1R}.
}
\examples{
RBF_Ftest(27.0, c(3, 48), 52, 3.2, c(3, 33), 37)
}
\references{
\itemize{
    \item \insertRef{Verhagen2014}{ReplicationBF}
    \item \insertRef{Harms2016}{ReplicationBF}
}
}
