% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF_tTest.R
\name{RBF_ttest}
\alias{RBF_ttest}
\title{Calculate a Replication Bayes Factor for a two-sample t-Test.}
\usage{
RBF_ttest(t.orig, n.orig, t.rep, n.rep, method = "NormApprox",
  M = 1e+05, store.samples = FALSE)
}
\arguments{
\item{t.orig}{t-statistic of the original study.}

\item{n.orig}{Numeric vector (2 elements) with cell sizes.}

\item{t.rep}{t-tstatistic of the replication study.}

\item{n.rep}{Numeric vector (2 elements) with cell sizes.}

\item{method}{Either \code{"NormApprox"} for a Normal approximation of the
original study's posterior distribution or \code{"MCMC"} for an
approximation through the Metropolis algorithm.}

\item{M}{Number of posterior samples (either MCMC or Normal distribution).}

\item{store.samples}{If \code{TRUE}, the returned object contains \code{M}
samples from the original study's posterior distribution (required for
\code{plot_RBF}).}
}
\value{
An object of \code{ReplicationBF} class.
}
\description{
\code{RBF_ttest} calculates a Replication Bayes Factor (Verhagen &
Wagenmakers, 2014) for t-Tests.
}
\details{
The Replication Bayes Factor is a marginal likelihood ratio between two
models, characterized by the following positions:
\itemize{
    \item \strong{H0}: The position of a skeptic, who does not place
        confidence in the findings of the original study and assumes
        \eqn{\theta \approx 0}.
    \item \strong{Hr}: The position of a proponent, who expects the original
        study to be a faithful estimation of the effect size and for which
        the replication offers additional evidence. This is formulated as
        \eqn{\theta \approx \theta_{orig}} with \eqn{\theta_{orig}} being
        the effect size estimate from the original study.
}

In contrast to the originally proposed Replication Bayes factor, this method
estimates the Bayes factor using importance sampling. This yields different
results than the code provided by Verhagen and Wagenmakers (2014), but is
more stable and less biased than the Monte Carlo estimate (see Bos, 2002) in
cases where original and replication study yield different effect size
estimates.
}
\examples{
\dontrun{
# Example 1 from Verhagen & Wagenmakers (2014)
# Using a Normal approximation to the original's posterior distribution
RBF_ttest(2.18, c(10, 11), 3.06, c(27, 27), method = "NormApprox")
RBF_ttest(2.18, c(10, 11), 0.25, c(27, 27), method = "NormApprox")
RBF_ttest(2.18, c(10, 11), 2.44, c(16, 17), method = "NormApprox")

# Using MCMC to draw samples from the original's posterior distribution
RBF_ttest(2.10, c(11, 11), 3.06, c(27, 28), method = "MCMC")
RBF_ttest(2.18, c(10, 11), 0.25, c(27, 27), method = "MCMC")
RBF_ttest(2.18, c(10, 11), 2.44, c(16, 17), method = "MCMC")
}
}
\references{
\itemize{
    \item \insertRef{Verhagen2014}{ReplicationBF}
    \item \insertRef{Bos2002}{ReplicationBF}
}
}
