% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaClasses.R
\docType{class}
\name{ReplicationBF-class}
\alias{ReplicationBF-class}
\title{General S4 class for a single Replication Bayes Factor}
\description{
General S4 class for a single Replication Bayes Factor
}
\section{Slots}{

\describe{
\item{\code{bayesFactor}}{The numerical value of the Replication Bayes Factor.}

\item{\code{approxMethod}}{Method used to approximate the marginal likelihoods.}

\item{\code{posteriorSamplesOriginal}}{A \code{data.frame} containing the samples
from the original study's posterior distribution.}

\item{\code{posteriorSamplesReplication}}{A \code{data.frame} containing samples
from the replication study's posterior distribution.}

\item{\code{test}}{A string that contains which test was used.}

\item{\code{functionCall}}{String containing the function call.}

\item{\code{originalStudy}}{Data from the original study that went into the analysis.}

\item{\code{replicationStudy}}{Data from the replication study that was used for the
calculation of the Bayes factor.}
}}

