% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplicationBF.R
\docType{package}
\name{ReplicationBF}
\alias{ReplicationBF}
\alias{ReplicationBF-package}
\title{ReplicationBF: A package to calculate Replication Bayes Factors.}
\description{
The ReplicationBF package provides functions for two common statistical
scenarios: t- and F-tests for difference in one or more group means.
}
