% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RBF.R
\name{plot_RBF}
\alias{plot_RBF}
\title{Plot \code{ReplicationBF} object}
\usage{
plot_RBF(rbf.object, use.ggplot = TRUE)
}
\arguments{
\item{rbf.object}{Object from an \code{RBF_*} call. Has to include Posterior
samples (\code{store.samples = TRUE}).}

\item{use.ggplot}{If \code{TRUE} uses \code{ggplot2}. \code{FALSE} not yet
supported.}
}
\value{
A \code{ggplot} object containing the plot.
}
\description{
This method can be used to visually represent a Replication Bayes Factor.
}
\details{
\code{plot_RBF} generally is a Savage-Dickey ratio representation:
Prior and Posterior distributions are plotted and the ratio of the
distributions at \eqn{\delta = 0} is calculated (see Wagenmakers et al., 2010).
}
\references{
\itemize{
    \item \insertRef{Wagenmakers2010}{ReplicationBF}
}
}
